serverless-framework/extract-pdftables/pdf-table-extract/requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
distro==1.7.0
idna==3.3
numpy==1.22.4
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
tabula-py==2.3.0
urllib3==1.26.9